add test alert bar
